third_party/blink/web_tests/http/tests/security/xss-DENIED-window-name-alert-expected.txt

CONSOLE MESSAGE: SecurityError: Failed to read a named property 'alert' from 'Window': Blocked a frame with origin "null" from accessing a cross-origin frame.
